requirements.txt
# Python 3.9-compatible core stack for the pipeline (SSD itself is installed separately)
# Pin numpy<2 to avoid incompatibilities with older RL/vision deps on some setups.
numpy>=1.21,<2
pandas>=1.3
matplotlib>=3.5
pyyaml>=6.0
tqdm>=4.60
torch>=2.0
tensorboard>=2.12
